* removed neighbour selectors (not really needed at all). Improves compilation speed, memory usage and binary size! * removed clustering specific code in favour of more generic property code * added GetEdgeHistogram() * renamed some functions/options ("degree"->"vertex")
git-svn-id: https://svn.forked.de/graph-tool/trunk@9 d4600afd-f417-0410-95de-beed9576f240

+++ src/graph/graph_selectors.hh
@@ -104,7 +104,7 @@ struct scalarS
     scalarS(){}
     scalarS(std::string scalar_property, const GraphInterface& g): 
 	_scalar_property(scalar_property), _g(&g) {}
-    typedef boost::mpl::vector<long double,double,float,long,unsigned long,int,unsigned int,short,unsigned short,char,unsigned char,bool,std::string> scalar_types;
+    typedef boost::mpl::vector<double,long double,float,long,unsigned long,int,unsigned int,short,unsigned short,char,unsigned char,bool,std::string> scalar_types;
     template <class Graph, class VertexOrEdge> 
     double operator()(const VertexOrEdge& v, const Graph &g) const 
     {
